requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
emoji==2.15.0
faiss-cpu==1.12.0
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.63
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.1.147
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==1.26.4
olefile==0.47
ollama==0.6.0
orjson==3.11.4
packaging==24.2
pillow==12.0.0
propcache==0.4.1
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pypdf==6.2.0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.35
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.18
unstructured-client==0.42.3
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==2.0.1
yarl==1.22.0
